feat: refac icons with hoc
